requirements.txt
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
whisper-timestamped==1.15.4
openai-whisper==20231105
dtw-python==1.5.3
Flask==3.1.0
coqui-tts==0.26.0
gunicorn